+++ components/Layout.js
@@ -7,7 +7,7 @@ export default function Layout({ children }) {
   return (
     <div className="bg-transparent flex flex-row justify-start w-screen h-screen overflow-auto">
         <Navbar/>
-        <div className="flex flex-col">
+        <div className="flex flex-col h-full w-full">
           {children}
         </div>
     </div>
