+++ package.json
@@ -16,6 +16,8 @@
     "@fortawesome/free-regular-svg-icons": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "@splinetool/react-spline": "^2.2.1",
+    "@splinetool/runtime": "^0.9.73",
     "next": "12.2.0",
     "next-themes": "^0.2.0",
     "react": "18.2.0",
